refactored into controllers and partials
