+++ Gemfile
@@ -2,21 +2,21 @@ source 'https://rubygems.org'
 ruby '2.2.2'
 
 gem 'bcrypt'
-gem 'database_cleaner'
 gem 'data_mapper'
 gem 'dm-postgres-adapter'
 gem 'dm-timestamps'
-# gem 'rspec-sinatra'
+gem 'rake'
 gem 'sinatra'
 gem 'sinatra-flash'
+gem 'sinatra-partial'
 
 group :development, :test do
-  gem 'capybara'
-  gem 'cucumber'
-  gem 'rspec'
-  gem 'rubocop'
-  gem 'rubocop-rspec'
-  gem 'coveralls', require: false
+	gem 'capybara'
+	gem 'database_cleaner'
+	gem 'rspec'
+	gem 'rubocop'
+	gem 'rubocop-rspec'
+	gem 'coveralls', require: false
 end
 
 gem "codeclimate-test-reporter", group: :test, require: nil
