+++ app/controllers/sessions.rb
@@ -1,0 +1,24 @@
+class Chitter_Challenge < Sinatra::Base
+
+ get '/sessions/new' do
+	erb :'sessions/new'
+ end
+
+ post '/sessions' do
+	user = User.authenticate(params[:email], params[:password])
+	if user
+		session[:user_id] = user.id
+		redirect to('/peeps')
+	else
+		flash.now[:errors] = ['The email or password is incorrect']
+		erb :'sessions/new'
+	end
+ end
+
+ delete '/sessions' do
+	session[:user_id] = nil
+	flash.keep[:notice] = 'Thanks for visiting, come back soon'
+	redirect to('/sessions/new')
+  erb :'sessions/new'
+ end
+end
