+++ app/views/_flash.erb
@@ -1,0 +1,16 @@
+        <% if flash[:errors] && !flash[:errors].empty? %>
+	    Please refer to the following errors below:
+	    <ul id='errors'>
+        <% flash[:errors].each do |error| %>
+		      <li><%= error %></li>
+      <% end %>
+      </ul>
+    <% end %>
+
+    <% if flash[:notice] %>
+      <div id='notice'>
+      <%= flash[:notice] %>
+      </div>
+    <% end %>
+
+
