+++ app/views/_welcome.erb
@@ -1,0 +1,7 @@
+    <% if current_user %>
+	    Welcome, <%= current_user.email %>
+	    <form method='POST' action='/sessions'>
+		    <input type='hidden' name='_method' value='delete'>
+		    <input type='submit' value='Sign Out'>
+		  </form>
+    <% end %>
