+++ app/views/layout.erb
@@ -5,29 +5,8 @@
   </head>
 
   <body>
-    <% if current_user %>
-	    Welcome, <%= current_user.email %>
-	    <form method='POST' action='/sessions'>
-		    <input type='hidden' name='_method' value='delete'>
-		    <input type='submit' value='Sign Out'>
-		  </form>
-    <% end %>
-
-    <% if flash[:errors] && !flash[:errors].empty? %>
-	    Please refer to the following errors below:
-	    <ul id='errors'>
-        <% flash[:errors].each do |error| %>
-		      <li><%= error %></li>
-      <% end %>
-      </ul>
-    <% end %>
-
-    <% if flash[:notice] %>
-      <div id='notice'>
-      <%= flash[:notice] %>
-      </div>
-    <% end %>
-
-    <%= yield %>
+	  <%= partial :welcome %>
+	  <%= partial :flash %>
+	  <%= yield %>
   </body>
 </html>
