+++ app/views/sessions/new.erb
@@ -2,10 +2,10 @@
 
 <form action='/sessions' method='POST'>
 	<label for='email'>
-		Email: <input name='email' type='email' placeholder='email'>
+		Email: <input name='email' type='email' placeholder='Email'>
 	</label>
 	<label for='password'>
-		Password: <input name='password' type='password' placeholder='password'>
+		Password: <input name='password' type='password' placeholder='Password'>
 	</label>
 	<input type='submit' value='Sign In'>
 </form>
